refactor: extract AnalysisWorkerQueue into analysis_queue module
- Move AnalysisWorkerQueue, AnalysisJobRequest, AnalysisQueueState
  and related constants/helpers from state.rs into analysis_queue.rs
- Reduces state.rs by ~260 lines (1758 → ~1500)
- First step in AppState decomposition
