+++ crates/codelens-mcp/src/main.rs
@@ -1,3 +1,4 @@
+mod analysis_queue;
 mod authority;
 mod dispatch;
 mod error;
