+++ crates/codelens-mcp/src/state.rs
@@ -5,8 +5,12 @@ use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 use std::fs;
 use std::path::{Path, PathBuf};
-use std::sync::{Arc, Condvar, Mutex, OnceLock};
+use std::sync::{Arc, Mutex, OnceLock};
 
+use crate::analysis_queue::{
+    analysis_job_cost_units, AnalysisJobRequest, AnalysisWorkerQueue, HTTP_ANALYSIS_WORKER_COUNT,
+    STDIO_ANALYSIS_WORKER_COUNT,
+};
 use crate::error::CodeLensError;
 use crate::telemetry::ToolMetricsRegistry;
 use crate::tool_defs::{ToolPreset, ToolProfile, ToolSurface};
@@ -15,9 +19,6 @@ use std::collections::VecDeque;
 
 const MAX_ANALYSIS_ARTIFACTS: usize = 64;
 const MAX_ANALYSIS_JOBS: usize = 128;
-const MAX_PENDING_ANALYSIS_REQUESTS: usize = 32;
-const STDIO_ANALYSIS_WORKER_COUNT: usize = 1;
-const HTTP_ANALYSIS_WORKER_COUNT: usize = 2;
 const ANALYSIS_TTL_MS: u64 = 24 * 60 * 60 * 1000;
 const WATCHER_RECENT_FAILURE_WINDOW_SECS: i64 = 15 * 60;
 pub(crate) const PREFLIGHT_TTL_MS: u64 = 10 * 60 * 1000;
@@ -28,19 +29,6 @@ fn default_risk_level() -> String {
 
 fn default_verifier_status() -> String {
     "caution".to_owned()
-}
-
-fn analysis_job_cost_units(kind: &str) -> usize {
-    match kind {
-        "impact_report" => 1,
-        "refactor_safety_report" => 2,
-        "dead_code_report" => 3,
-        _ => 2,
-    }
-}
-
-fn queue_pending_cost_units(pending: &VecDeque<AnalysisJobRequest>) -> usize {
-    pending.iter().map(|request| request.cost_units).sum()
 }
 
 fn push_unique_string(items: &mut Vec<String>, value: String) {
@@ -63,24 +51,6 @@ fn normalize_path_for_project(project_root: &Path, path: &str) -> String {
         .replace('\\', "/")
 }
 
-struct AnalysisJobRequest {
-    job_id: String,
-    kind: String,
-    arguments: Value,
-    profile_hint: Option<String>,
-    cost_units: usize,
-}
-
-struct AnalysisQueueState {
-    pending: VecDeque<AnalysisJobRequest>,
-    active_jobs: usize,
-    active_cost_units: usize,
-}
-
-struct AnalysisWorkerQueue {
-    inner: Arc<(Mutex<AnalysisQueueState>, Condvar)>,
-}
-
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub(crate) enum RuntimeTransportMode {
     Stdio,
@@ -95,7 +65,7 @@ impl RuntimeTransportMode {
         }
     }
 
-    fn as_str(&self) -> &'static str {
+    pub(crate) fn as_str(&self) -> &'static str {
         match self {
             Self::Stdio => "stdio",
             Self::Http => "http",
@@ -135,126 +105,6 @@ impl RuntimeDaemonMode {
             Self::ReadOnly => "read-only",
             Self::MutationEnabled => "mutation-enabled",
         }
-    }
-}
-
-impl AnalysisWorkerQueue {
-    fn new(state: &AppState) -> Self {
-        let inner = Arc::new((
-            Mutex::new(AnalysisQueueState {
-                pending: VecDeque::new(),
-                active_jobs: 0,
-                active_cost_units: 0,
-            }),
-            Condvar::new(),
-        ));
-        let worker_limit = state.analysis_worker_limit();
-        state.metrics().record_analysis_worker_pool(
-            worker_limit,
-            state.analysis_cost_budget(),
-            state.transport_mode().as_str(),
-        );
-        for _ in 0..worker_limit {
-            let inner_clone = Arc::clone(&inner);
-            let worker_state = state.clone_for_worker();
-            std::thread::spawn(move || loop {
-                let request = {
-                    let (lock, condvar) = &*inner_clone;
-                    let mut guard = lock.lock().unwrap_or_else(|p| p.into_inner());
-                    loop {
-                        if guard.pending.is_empty() {
-                            guard = condvar.wait(guard).unwrap_or_else(|p| p.into_inner());
-                            continue;
-                        }
-                        let cost_budget = worker_state.analysis_cost_budget();
-                        let next_index = guard.pending.iter().position(|request| {
-                            let allowed_parallelism = worker_state
-                                .analysis_parallelism_for_profile(request.profile_hint.as_deref());
-                            guard.active_jobs < allowed_parallelism
-                                && guard.active_cost_units + request.cost_units <= cost_budget
-                        });
-                        if let Some(index) = next_index {
-                            let request = guard.pending.remove(index);
-                            guard.active_jobs += 1;
-                            if let Some(request) = request.as_ref() {
-                                guard.active_cost_units += request.cost_units;
-                            }
-                            let remaining_depth = guard.pending.len();
-                            let remaining_cost_units =
-                                queue_pending_cost_units(&guard.pending) + guard.active_cost_units;
-                            break request
-                                .map(|request| (request, remaining_depth, remaining_cost_units));
-                        }
-                        guard = condvar.wait(guard).unwrap_or_else(|p| p.into_inner());
-                    }
-                };
-                if let Some((request, remaining_depth, remaining_cost_units)) = request {
-                    if worker_state
-                        .get_analysis_job(&request.job_id)
-                        .as_ref()
-                        .map(|job| job.status.as_str())
-                        == Some("cancelled")
-                    {
-                        let (lock, condvar) = &*inner_clone;
-                        let mut guard = lock.lock().unwrap_or_else(|p| p.into_inner());
-                        guard.active_jobs = guard.active_jobs.saturating_sub(1);
-                        guard.active_cost_units =
-                            guard.active_cost_units.saturating_sub(request.cost_units);
-                        condvar.notify_all();
-                        continue;
-                    }
-                    let request_cost = request.cost_units;
-                    worker_state
-                        .metrics()
-                        .record_analysis_job_started(remaining_depth, remaining_cost_units);
-                    let final_status = crate::tools::reports::run_analysis_job_from_queue(
-                        &worker_state,
-                        request.job_id,
-                        request.kind,
-                        request.arguments,
-                    );
-                    let (remaining_depth, remaining_cost_units) = {
-                        let (lock, condvar) = &*inner_clone;
-                        let mut guard = lock.lock().unwrap_or_else(|p| p.into_inner());
-                        guard.active_jobs = guard.active_jobs.saturating_sub(1);
-                        guard.active_cost_units =
-                            guard.active_cost_units.saturating_sub(request_cost);
-                        let remaining_depth = guard.pending.len();
-                        let remaining_cost_units =
-                            queue_pending_cost_units(&guard.pending) + guard.active_cost_units;
-                        condvar.notify_all();
-                        (remaining_depth, remaining_cost_units)
-                    };
-                    worker_state.metrics().record_analysis_job_finished(
-                        final_status,
-                        remaining_depth,
-                        remaining_cost_units,
-                    );
-                }
-            });
-        }
-        Self { inner }
-    }
-
-    fn enqueue(&self, request: AnalysisJobRequest) -> Result<(usize, usize, bool), CodeLensError> {
-        let (lock, condvar) = &*self.inner;
-        let mut guard = lock.lock().unwrap_or_else(|p| p.into_inner());
-        if guard.pending.len() >= MAX_PENDING_ANALYSIS_REQUESTS {
-            return Err(CodeLensError::Validation(format!(
-                "analysis queue is full (>{MAX_PENDING_ANALYSIS_REQUESTS} pending jobs)"
-            )));
-        }
-        let insert_at = guard
-            .pending
-            .iter()
-            .position(|existing| existing.cost_units > request.cost_units)
-            .unwrap_or(guard.pending.len());
-        let priority_promoted = insert_at < guard.pending.len();
-        guard.pending.insert(insert_at, request);
-        let depth = guard.pending.len();
-        let weighted_depth = queue_pending_cost_units(&guard.pending) + guard.active_cost_units;
-        condvar.notify_all();
-        Ok((depth, weighted_depth, priority_promoted))
     }
 }
